Actually use PSP_getMsgBuf() to implement PSI_getReservation() and PSI_getSlots()
git-svn-id: http://svn.ipd.uka.de/repos/ParaStation/trunk/psmgmt@9061 21b3a009-d9f9-44df-8eaf-6b729cfaa075

+++ include/psipartition.h
@@ -374,7 +374,7 @@ PSrsrvtn_ID_t PSI_getReservation(uint32_t nMin, uint32_t nMax, uint16_t tpp,
  *
  * @see PSI_getReservation()
  */
-int PSI_getSlots(uint32_t num, PSrsrvtn_ID_t resID, PSnodes_ID_t *nodes);
+int PSI_getSlots(uint16_t num, PSrsrvtn_ID_t resID, PSnodes_ID_t *nodes);
 
 /**
  * @brief Get node to spawn process to.
